% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allofus-package.R
\docType{package}
\name{allofus-package}
\alias{allofus}
\alias{allofus-package}
\title{allofus: Interface for 'All of Us' Researcher Workbench}
\description{
Streamline use of the 'All of Us' Researcher Workbench (\url{https://www.researchallofus.org/data-tools/workbench/})with tools to extract and manipulate data from the 'All of Us' database. Increase interoperability with the Observational Health Data Science and Informatics ('OHDSI') tool stack by decreasing reliance of 'All of Us' tools and allowing for cohort creation via 'Atlas'. Improve reproducible and transparent research using 'All of Us'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://roux-ohdsi.github.io/allofus/}
  \item \url{https://github.com/roux-ohdsi/allofus}
  \item Report bugs at \url{https://github.com/roux-ohdsi/allofus/issues}
}

}
\author{
\strong{Maintainer}: Louisa Smith \email{l.smith@northeastern.edu} (\href{https://orcid.org/0000-0001-9029-4644}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rob Cavanaugh \email{rcavanaugh1@mghihp.edu} (\href{https://orcid.org/0000-0002-2114-6565}{ORCID}) [copyright holder]
}

}
\keyword{internal}
