% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_bucket_to_workspace.R
\name{aou_bucket_to_workspace}
\alias{aou_bucket_to_workspace}
\title{Move files from a bucket to your workspace}
\usage{
aou_bucket_to_workspace(files, bucket_name = Sys.getenv("WORKSPACE_BUCKET"))
}
\arguments{
\item{files}{The name of a file in your bucket or a vector of multiple files.}

\item{bucket_name}{Name of your bucket. Recommend leaving the default}
}
\description{
This step retrieves a file you have saved permanently in your bucket
into your workspace where you can read it into R using a function like write.csv().
}
