% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aou_cohort_example}
\alias{aou_cohort_example}
\title{Example cohort definition and SQL query from ATLAS}
\format{
\code{aou_cohort_example}
\describe{
\item{cd}{Cohort definition object}
\item{cd_sql}{SQL query for cohort definition}
}
}
\usage{
aou_cohort_example
}
\description{
A list containing a sample cohort definition and SQL query from the demo ATLAS for a cohort of stroke survivors.
}
\keyword{datasets}
