% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_temp_tables.R
\name{aou_compute}
\alias{aou_compute}
\title{Compute a dplyr tbl SQL query into a temp table}
\usage{
aou_compute(.data, con = getOption("aou.default.con"))
}
\arguments{
\item{.data}{result of tbl(con, "table") \%>\% ... query}

\item{con}{connection from aou_connect(). Set by default}
}
\value{
a name to a temporary table in the database.
}
\description{
Computes a temporary table from a dplyr chain that returns an
SQL query (e.g., tbl(con, table)) and returns the name of the temporary table.
May be useful to create intermediate tables to reduce long queries.
The temporary table will only exist for the current session and will nee
to be created again a new session.
}
\examples{
\dontshow{if (on_workbench()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

con <- aou_connect()
tmp_tbl = tbl(con, "concept") \%>\%
   select(concept_id) \%>\%
   head(10) \%>\%
   aou_compute()

tbl(con, tmp_tbl)

\dontshow{\}) # examplesIf}
}
