% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aou_concept_codes}
\alias{aou_concept_codes}
\title{Concept codes and survey answers}
\format{
aou_concept_codes
\describe{
\item{code}{response from the observation table}
\item{answer}{Text responses}
}
}
\usage{
aou_concept_codes
}
\description{
A data frame containing concept codes (\code{code}) and text responses (\code{answer}) for
the SDOH and COPE surveys.
}
\keyword{datasets}
