% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_temp_tables.R
\name{aou_create_temp_table}
\alias{aou_create_temp_table}
\title{Creates a temporary table from a local data frame or tibble}
\usage{
aou_create_temp_table(df)
}
\arguments{
\item{df}{a local dataframe or tibble}
}
\value{
a name to a temporary table in the database.
}
\description{
Experimental function that builds a local tibble into an SQL query and
generates a temporary table. Tables
generally will need to be small in size (<1500 rows). The table will only
exist for the current connection session and will need to be created again
in a new session.
}
\examples{
\dontshow{if (on_workbench()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
con <- aou_connect()
df = tibble(
  category = c("DELIV", "ECT", "AB", "LB", "SB", "SA"),
  max_term = as.integer(c(301, 84, 168, 301, 301, 139)),
  min_term = as.integer(c(140, 42, 42, 161, 140, 28)),
  retry =    as.integer(c(28, 14, 14, 28, 28, 14))
)
tmp_tbl = aou_create_temp_table(df)
tbl(con, "tmp_tbl")

\dontshow{\}) # examplesIf}
}
