% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_enrollment_date.R
\name{aou_enrollment_date}
\alias{aou_enrollment_date}
\title{Allofus Enrollment Date}
\usage{
aou_enrollment_date(
  cohort = NULL,
  con = getOption("aou.default.con"),
  collect = FALSE,
  ...
)
}
\arguments{
\item{cohort}{An SQL tbl with a person_id column. Does not currently work with dataframes.}

\item{con}{dbplyr connection to the aou bigquery database.}

\item{collect}{whether to collect the results into a local data frame. Default is FALSE.}

\item{...}{not used}
}
\value{
a SQL query or local data frame
}
\description{
Calculates the date of the first consent PII observation for each person in the cohort.
}
\details{
Implements query suggested by allofus described here: https://support.researchallofus.org/hc/en-us/articles/13176125767188-How-to-find-participant-enrollment-data
}
\examples{
\dontshow{if (on_workbench()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

con <- aou_connect()
cohort <- dplyr::tbl(con, "person") \%>\%
  dplyr::filter(person_id > 5000000) \%>\%
  dplyr::select(person_id)

aou_enrollment_date(
  cohort
)
\dontshow{\}) # examplesIf}
}
