% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_pull_concepts.R
\name{aou_get_concepts}
\alias{aou_get_concepts}
\title{Get concepts from a specified domain}
\usage{
aou_get_concepts(
  ...,
  domain = c("condition", "measurement", "observation", "procedure", "drug", "device")
)
}
\arguments{
\item{...}{Arguments passed to specific `aou_get_{}_concepts()` function}

\item{domain}{A character string specifying the domain to retrieve concepts from. Must be one of: "condition", "measurement", "observation", "procedure", "drug", "device"}
}
\value{
A list of concepts from the specified domain
}
\description{
Get concepts from a specified domain
}
\examples{
aou_get_concepts(cohort, c(123, 456), start_date, end_date, domain = "condition")

}
