% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_pull_concepts.R
\name{aou_get_measurement_concepts}
\alias{aou_get_measurement_concepts}
\title{Get concepts from the measurement table for a given cohort}
\usage{
aou_get_measurement_concepts(cohort, concepts, start_date, end_date, ...)
}
\arguments{
\item{cohort}{A cohort object}

\item{concepts}{A vector of concept IDs to filter by}

\item{start_date}{The name of the column with the start date for filtering measurements}

\item{end_date}{The name of the column with the end date for filtering measurements}

\item{...}{Additional arguments to pass to aou_join}
}
\value{
A data frame with columns person_id, date, concept_id, concept_name, and domain = "measurement"
}
\description{
Get concepts from the measurement table for a given cohort
}
\examples{
aou_get_measurement_concepts(cohort, c(123, 456), start_date, end_date)
}
