% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_ls.R
\name{aou_ls_workspace}
\alias{aou_ls_workspace}
\title{List the current files in your workspace.}
\usage{
aou_ls_workspace(pattern = "*.csv", silent = FALSE)
}
\arguments{
\item{pattern}{pattern like *.csv or a single file name e.g., mydata.csv}

\item{silent}{logical; whether to print the files found}
}
\description{
List the current files in your workspace.
}
