% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_pull_concepts.R
\name{aou_pull_survey_concepts}
\alias{aou_pull_survey_concepts}
\title{Get survey questions from AoU for a given cohort}
\usage{
aou_pull_survey_concepts(
  cohort,
  concepts,
  collect = TRUE,
  reshape = FALSE,
  ...
)
}
\arguments{
\item{cohort}{tbl; reference to a table with a column called "person_id"}

\item{concepts}{num; a vector of concept ids for questions in the survey table}

\item{collect}{lgl; whether to collect from the database}

\item{reshape}{lgl; whether to turn the long data into wide data with clean variable names}
}
\value{
if reshape = FALSE, a dataframe or remote tbl with columns person_id, date (survey_datetime),
 concept_id (question_concept_id), question, answer. If reshape = TRUE, a dataframe with
 questions as columns. Those with multiple answers per person ("checkbox" questions) are list-columns.
}
\description{
Get survey questions from AoU for a given cohort
}
\examples{
\dontrun{
survey_data <- aou_pull_survey_concepts(cohort, concepts = c(1157, 124839), reshape = TRUE)
 )}

}
