% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_ls.R
\name{aou_read_from_bucket}
\alias{aou_read_from_bucket}
\title{Read the contents of a file in your bucket into memory.}
\usage{
aou_read_from_bucket(
  filename,
  silent = FALSE,
  bucket = Sys.getenv("WORKSPACE_BUCKET"),
  project = Sys.getenv("GOOGLE_PROJECT"),
  read_fn = readr::read_table,
  gsutil_args = "",
  ...
)
}
\arguments{
\item{filename}{The name of the file to read in. This can either be a full bucket path,
or the path to the file relative to the bucket associated with the workspace.}

\item{bucket}{The bucket to read the file from, if not specified in the filename.}

\item{project}{The Google Cloud project to use.}

\item{read_fn}{The function to use to read the file. Defaults to `readr::read_table`.}

\item{gsutil_args}{Additional arguments to pass to `gsutil cat`.}

\item{...}{Other arguments passed to `read_fn`.}
}
\value{
The contents of the file as returned by read_fn.
}
\description{
Streams the contents of a file in your bucket into memory.
}
\examples{
on_workbench()
files = aou_ls_bucket("*.tsv")
file = files[1]
aou_read_from_bucket(file, read_fn=readr::read_tsv)

}
