% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_session_info.R
\name{aou_session_info}
\alias{aou_session_info}
\title{Print session information for the AoU R environment}
\usage{
aou_session_info(CDR = getOption("aou.default.cdr"))
}
\arguments{
\item{CDR}{The name of the CDR to use. Defaults to `getOption("aou.default.cdr")`}
}
\value{
A list with three elements: the platform, the AoU release, and the packages
}
\description{
Print session information for the AoU R environment
}
\examples{
\dontrun{
  con <- allofus::aou_connect()
  allofus::aou_session_info()
}
}
