% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{aou_survey}
\alias{aou_survey}
\title{Funciton to query observation table for survey responses}
\usage{
aou_survey(
  con,
  cohort,
  questions,
  collect = TRUE,
  answer_output = c("value", "concept_id"),
  question_output = c("value", "concept_id")
)
}
\arguments{
\item{con}{connection to the allofus SQL database.}

\item{cohort}{tbl or dataframe with a cohort that includes a column called person_id}

\item{questions}{either a vector of concept_ids or concept_codes for questions to return results}

\item{collect}{whether to return the results as a local (TRUE) or database table}

\item{answer_output}{whether to return the survey responses in their text format (value) or concept_id}

\item{question_output}{hether to return the survey questions (columns) in their text format (value) or concept_id}
}
\description{
Funciton to query observation table for survey responses
}
\examples{
\dontrun{
con <- aou_connect()
cohort = tbl(con, "person") \%>\% filter(person_id > 5000000) \%>\% select(person_id, year_of_birth, gender_concept_id)
aou_survey(con = con,
cohort = cohort,
questions = c(1585375, 1586135),
answer_output = "value",
question_output = "value")
}
}
