% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_connect.R
\name{aou_test_connect}
\alias{aou_test_connect}
\title{Internal function to test allofus package on fake database}
\usage{
aou_test_connect(
  cache = TRUE,
  cache_dir = Sys.getenv("AOU_CACHE_DIR"),
  overwrite = FALSE
)
}
\arguments{
\item{cache}{logical indicating whether to cache the downloaded data}

\item{cache_dir}{character string specifying the directory to store cached data}

\item{overwrite}{logical indicating whether to overwrite existing cached data}
}
\value{
a database connection object
}
\description{
Internal function to test allofus package on fake database
}
\details{
This function downloads a test OMOP database from the OHDSI Eunomia package, renames tables and fields to lowercase,
removes tables that are not relevant to All of Us, adds missing fields to relevant tables, and drops fields that are not relevant to All of Us.
It then tests the connection to the database by checking if the person_id column is in the person table.
}
\examples{
\dontrun{
con <- aou_test_connect()
}
}
