% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_test.R
\name{aou_test_package}
\alias{aou_test_package}
\title{Function to test all of us package on AllofUs Researcher Workbench using documented examples}
\usage{
aou_test_package()
}
\value{
result of test
}
\description{
Function to test all of us package on AllofUs Researcher Workbench using documented examples
}
