% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_ls.R
\name{aou_workspace_to_bucket}
\alias{aou_workspace_to_bucket}
\title{Save a file from your workspace to your bucket}
\usage{
aou_workspace_to_bucket(
  file,
  dir = "",
  recursive = TRUE,
  bucket = getOption("aou.default.bucket")
)
}
\arguments{
\item{file}{The name of a file in your bucket, a vector of multiple files, a directory,
or a file pattern (e.g. ".csv"). See Details.}

\item{dir}{Optional directory in the bucket to save files to.}

\item{recursive}{Whether to include subdirectories. Defaults to \code{TRUE}.}

\item{bucket}{Bucket to save files to. Defaults to \code{getOption("aou.default.bucket")},
which is \code{Sys.getenv('WORKSPACE_BUCKET')} unless specified otherwise.}
}
\description{
This function moves a file saved in a workspace
to a bucket, where it can be retrieved even if the environment is deleted. To use, first save the desired
object as a file to the workspace (e.g., \code{write.csv(object, "filename.csv")}) and then run this function
(e.g., \code{aou_workspace_to_bucket(files = "filename.csv")}). See \url{https://cloud.google.com/storage/docs/gsutil/commands/cp} for details on the
underlying function.
}
\examples{
\dontrun{
aou_workspace_to_bucket("data1.csv")
# a file to a specific directory in the bucket
aou_workspace_to_bucket("data2.csv", dir = "data")
# all of the files in this directory
aou_workspace_to_bucket("data/")
# multiple specific files
aou_workspace_to_bucket(c("data1.csv", "data2.csv"))
}

}
