% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_functions.R
\name{CAST}
\alias{CAST}
\title{Date add function to be passed along to SQL}
\usage{
CAST(...)
}
\arguments{
\item{...}{additional arguments passed to dbplyr SQL}
}
\value{
Internal, passed along to dbplyr SQL.
}
\description{
Date add function to be passed along to SQL
}
