% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_connection}
\alias{check_connection}
\title{Checks whether a connection is valid for functions. Internal}
\usage{
check_connection(con = getOption("aou.default.con"))
}
\arguments{
\item{con}{connection to aou database}
}
\description{
Checks whether a connection is valid for functions. Internal
}
\keyword{internal}
