% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{collect.tbl_aou}
\alias{collect.tbl_aou}
\title{Collects the data from a tbl_aou object}
\usage{
collect.tbl_aou(tbl_obj, convert_int64 = TRUE, ...)
}
\arguments{
\item{tbl_obj}{The tbl_aou object to collect data from}

\item{...}{Additional arguments to be passed to the \code{collect} function}
}
\value{
A data frame with the collected data
}
\description{
This function collects the data from a tbl_aou object and converts it to a regular data frame.
Finally, it converts any columns of type \code{integer64} to \code{double} for ease of use.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
