% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{compute.tbl_aou}
\alias{compute.tbl_aou}
\title{Compute function for tbl_aou objects}
\usage{
compute.tbl_aou(x, con = getOption("aou.default.con"))
}
\arguments{
\item{x}{The tbl_aou object to compute.}

\item{con}{The database connection to use. Defaults to the value of 'aou.default.con' option.}
}
\value{
A tbl.src_aou object representing the computed table.
}
\description{
This function takes a tbl_aou object and computes the SQL query associated with it.
It returns a tbl.src_aou object representing the computed table.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
