% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_functions.R
\name{date_add}
\alias{date_add}
\title{Date add function to be passed along to SQL}
\usage{
date_add(...)
}
\arguments{
\item{...}{additional arguments passed to dbplyr SQL}
}
\value{
Internal, passed along to dbplyr SQL.
}
\description{
Date add function to be passed along to SQL
}
