% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_pull_concepts.R
\name{get_domain_concepts}
\alias{get_domain_concepts}
\title{Retrieves domain concepts for a given cohort and time range}
\usage{
get_domain_concepts(
  cohort,
  concepts,
  start_date,
  end_date,
  tbl_name,
  date_column,
  concept_id_column,
  ...
)
}
\arguments{
\item{cohort}{A data frame containing person IDs}

\item{concepts}{A vector of concept IDs to retrieve}

\item{start_date}{The start date of the time range to retrieve concepts for}

\item{end_date}{The end date of the time range to retrieve concepts for}

\item{tbl_name}{The name of the table containing the domain concepts}

\item{date_column}{The name of the column containing the concept dates}

\item{concept_id_column}{The name of the column containing the concept IDs}

\item{...}{Additional arguments}
}
\description{
Retrieves domain concepts for a given cohort and time range
}
