% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_connect.R
\name{on_workbench}
\alias{on_workbench}
\title{Check to see whether you are on the All of Us workbench}
\usage{
on_workbench()
}
\value{
TRUE if you are on the workbench, FALSE otherwise
}
\description{
Use this function to check whether you are on the All of Us
Researcher Workbench. This is useful for writing code that can be used both
on the workbench and locally.
}
\examples{
on_workbench()
}
\keyword{internal}
