% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{read_aou_env}
\alias{read_aou_env}
\title{Read cached workspace environment variables from \verb{~/.aou-env}}
\usage{
read_aou_env()
}
\value{
A named character vector (empty if the file doesn't exist).
}
\description{
Read cached workspace environment variables from \verb{~/.aou-env}
}
\keyword{internal}
