% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tbl.src_aou}
\alias{tbl.src_aou}
\title{Create a custom tbl object from a data source}
\usage{
tbl.src_aou(src, from, ...)
}
\arguments{
\item{src}{The data source to create the tbl object from.}

\item{from}{The table to select from the data source.}

\item{...}{Additional arguments to be passed to the dplyr::tbl function.}
}
\value{
A reference to a remote table with class "tbl_aou".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Create a tbl object from a data source
con <- aou_connect()
person_table <- tbl(con, "person")

}
