% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_cohort_sql_df}
\alias{validate_cohort_sql_df}
\title{Checks whether a cohort is valid for functions. Internal. SQL cohorts only}
\usage{
validate_cohort_sql_df(cohort, con = getOption("aou.default.con"))
}
\arguments{
\item{cohort}{cohort provided to function}
}
\description{
Checks whether a cohort is valid for functions. Internal. SQL cohorts only
}
\keyword{internal}
