% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{workbench_cdr}
\alias{workbench_cdr}
\title{Resolve \code{WORKSPACE_CDR} from the Workbench CLI}
\usage{
workbench_cdr()
}
\value{
A \code{project.dataset} string, or \code{NULL} if it can't be determined.
}
\description{
Resolve \code{WORKSPACE_CDR} from the Workbench CLI
}
\keyword{internal}
