% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{workbench_google_project}
\alias{workbench_google_project}
\title{Resolve \code{GOOGLE_PROJECT} from the Workbench context file}
\usage{
workbench_google_project()
}
\value{
The workspace's Google project id, or \code{NULL} if it can't be determined.
}
\description{
Resolve \code{GOOGLE_PROJECT} from the Workbench context file
}
\keyword{internal}
