% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allofus-package.R
\docType{package}
\name{allofus-package}
\alias{allofus}
\alias{allofus-package}
\title{allofus: Tools for the AllofUs Resarcher Workbench}
\description{
Streamlining use of the 'AllofUs Researcher Workbench' notebooks and increasing interoperability with the OHDSI toolstack.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://roux-ohdsi.github.io/allofus/}
  \item \url{https://github.com/roux-ohdsi/allofus}
  \item Report bugs at \url{https://github.com/roux-ohdsi/allofus/issues}
}

}
\author{
\strong{Maintainer}: Rob Cavanaugh \email{r.cavanaugh@northeastern.edu} [copyright holder]

Authors:
\itemize{
  \item Louisa Smith \email{l.smith@northeastern.edu} [copyright holder]
}

}
\keyword{internal}
