% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allofus-package.R
\docType{package}
\name{allofus-package}
\alias{allofus}
\alias{allofus-package}
\title{allofus: R package to streamline use of the AllofUs Resarcher Workbench}
\description{
The allofus R package contains useful functions for working in the AllofUs Researcher Workbench.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://roux-ohdsi.github.io/allofus/}
  \item \url{https://github.com/roux-ohdsi/allofus}
  \item Report bugs at \url{https://github.com/roux-ohdsi/allofus/issues}
}

}
\author{
\strong{Maintainer}: Rob Cavanaugh \email{r.cavanaugh@northeastern.edu}

Authors:
\itemize{
  \item Louisa Smith \email{l.smith@northeastern.edu}
}

}
\keyword{internal}
