% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_atlas_cohort.R
\name{aou_atlas_cohort}
\alias{aou_atlas_cohort}
\title{Retrieve a cohort from ATLAS for use in AllofUs}
\usage{
aou_atlas_cohort(
  cohort_id,
  persistence_window = 548,
  end_date_buffer = 60,
  exclude_aou_visits = FALSE,
  base_url = "http://atlas-demo.ohdsi.org/WebAPI"
)
}
\arguments{
\item{cohort_id}{The ID of the cohort to retrieve}

\item{persistence_window}{The number of days to look back from the cohort start date for the observation period. Defaults to 548 days.}

\item{end_date_buffer}{The number of days to add to the end date of the observation period. Defaults to 60 days.}

\item{exclude_aou_visits}{Whether to exclude AllofUs-specific visit concepts from the observation period. Defaults to FALSE.}

\item{base_url}{The URL of the ATLAS instance to use ending in /WebAPI. Defaults to demo atlas at https://atlas-demo.ohdsi.org}
}
\value{
A dataframe with the resulting cohort. The SQL query used to generate the cohort is stored as an attribute.
}
\description{
This function retrieves a cohort definition from ATLAS and generates the cohort in All of Us.
Observation periods are first generated for each subject using the {aou_observation_period} function.
The resulting cohort is a dataframe with the cohort start and end dates for each subject.
The function is based on https://github.com/cmayer2/r4aou with some tweaks to generate the appropriate observation periods.
}
\examples{

\dontrun{
 cohort <- aou_atlas_cohort(1788061, base_url = "https://atlas-demo.ohdsi.org/WebAPI")
}

}
