% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_atlas_cohort.R
\name{aou_atlas_cohort}
\alias{aou_atlas_cohort}
\title{Retrieve a cohort from ATLAS for use in AllofUs}
\usage{
aou_atlas_cohort(
  cohort_id,
  persistence_window = 548,
  end_date_buffer = 60,
  exclude_aou_visits = FALSE,
  base_url = "http://api.ohdsi.org/WebAPI",
  collect = FALSE
)
}
\arguments{
\item{cohort_id}{The ID of the cohort to retrieve}

\item{base_url}{The URL of the ATLAS instance to use ending in /WebAPI. Defaults to demo atlas at https://atlas-demo.ohdsi.org}
}
\value{
a dataframe or list with the resulting cohort. see include_query
}
\description{
Retrieve a cohort from ATLAS for use in AllofUs
}
\examples{

\dontrun{
 con <- allofus::aou_connect()
 cohort <- aou_atlas_cohort(1788061, baseUrl = "https://atlas-demo.ohdsi.org/WebAPI")
}

}
