% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_ls.R
\name{aou_bucket_to_workspace}
\alias{aou_bucket_to_workspace}
\title{Move files from a bucket to your workspace}
\usage{
aou_bucket_to_workspace(
  file,
  dir = "",
  bucket = getOption("aou.default.bucket")
)
}
\arguments{
\item{file}{The name of a file in your bucket, a vector of multiple files, a directory,
or a file pattern (e.g. ".csv").}

\item{dir}{Optional directory in the workspace to save files to.}

\item{bucket}{Bucket to retrieve file from. Defaults to `getOption("aou.default.bucket")`,
which is `Sys.getenv('WORKSPACE_BUCKET')` unless specified otherwise.}
}
\value{
Nothing
}
\description{
This function retrieves a file from your bucket and moves it
into your workspace where it can be read into R, e.g., using a function like write.csv().
See https://cloud.google.com/storage/docs/gsutil/commands/cp for details on the
underlying function.
}
\examples{
\dontrun{
aou_bucket_to_workspace("data1.csv")
read.csv("data1.csv")
# a file to a specific directory in the workspace
aou_bucket_to_workspace("data2.csv", dir = "data")
read.csv("data/data2.csv")
# all of the files in this directory
aou_bucket_to_workspace("data/")
read.csv("data/data3.csv")
}

}
