% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aou_cohort_example}
\alias{aou_cohort_example}
\title{Example cohort definition and SQL query from ATLAS}
\format{
\code{aou_cohort_example}
\describe{
\item{cd}{Cohort definition object}
\item{cd_sql}{SQL query for cohort definition}
}
}
\usage{
aou_cohort_example
}
\description{
A data frame containing tables and columns in the All of Us
Controlled Tier Dataset v7 CDR Data Dictionary (C2022Q4R9) found here:
https://docs.google.com/spreadsheets/d/1XLVq84LLd0VZMioF2sPwyiaPw3EFp5c8o1CTWGPH-Yc/edit#gid=1815943286.
Note that the column 'value_source_value' has been manually added as it is
missing from the data dictionary.
}
\keyword{internal}
