% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_temp_tables.R
\name{aou_compute}
\alias{aou_compute}
\title{Compute a dplyr tbl SQL query into a temp table}
\usage{
aou_compute(.data, ..., con = getOption("aou.default.con"))
}
\arguments{
\item{.data}{result of tbl(con, "table") \%>\% ... query}

\item{...}{Other arugments passed to \code{bigrquery::bq_table_download()} when \code{collect = TRUE}}

\item{con}{connection from aou_connect(). Set by default}
}
\value{
a name to a temporary table in the database.
}
\description{
Experimental function that computes a temporary table from a tbl(con, "table")
dplyr chain that returns an SQL query (e.g., with show_query()). The
may be useful to create intermediate tables to reduce long queries.
It is a workaround for dplyr::compute(temporary = TRUE) which does not
currently work on the workbench. The table will only
exist for the current connection session and will need to be created again
in a new session.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontshow{if (on_workbench()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

con <- aou_connect()
tmp_tbl = tbl(con, "concept") \%>\%
   select(concept_id) \%>\%
   head(10) \%>\%
   aou_compute()

tbl(con, tmp_tbl)

\dontshow{\}) # examplesIf}
}
