% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_pull_concepts.R
\name{aou_concept_set}
\alias{aou_concept_set}
\title{Get occurrences of a concept set from AoU for a given cohort}
\usage{
aou_concept_set(
  cohort,
  concepts,
  start_date,
  end_date,
  domains = c("condition", "measurement", "observation", "procedure", "drug", "device"),
  return = "indicator",
  concept_set_name = "concept_set",
  min_n = 1,
  con = getOption("aou.default.con"),
  collect = TRUE,
  ...
)
}
\arguments{
\item{cohort}{tbl; reference to a table with a column called "person_id", and columns for start_date and end_date}

\item{concepts}{num; a vector of concept ids}

\item{start_date}{the name of the start_date column in the cohort table (unquoted)}

\item{end_date}{the name of the end_date column in the cohort table (unquoted)}

\item{return}{one of "indicator", "count", "all"; do you want to return a 1 if a person has any matching concepts and 0 if not ("indicator"),
the number of matching concepts per person ("count"), or all info about the matching concepts ("all"). Defaults to "indicator"}

\item{concept_set_name}{chr; If return = "indicator", name for that column. Defaults to "concept_set".}

\item{min_n}{dbl; If return = "indicator", the minimum number of occurrences per person to consider the indicator true. Defaults to 1.}

\item{con}{connection to the allofus SQL database. Defaults to getOption("aou.default.con"), which is set automatically if you use `aou_connect()`}

\item{collect}{lgl; whether to collect from the database}
}
\value{
a dataframe if collect = TRUE; a remote tbl if not
}
\description{
Get occurrences of a concept set from AoU for a given cohort
}
\examples{
\dontrun{
# simple example
aspirin_users <- pull_concepts(cohort, concepts = 1191, start_date = covariate_start_date,
 end_date = cohort_start_date, concept_set_name = "aspirin", domain = "drug")

# starting with person table
people <- tbl(con, "person") \%>\%
mutate(start_date = as.Date("1970-01-01"),
       end_date = as.Date("2023-05-24"))

dat <- aou_concept_set(cohort = people,
                         concepts = c(725115, 1612146, 1613031),
                         start_date = start_date,
                         end_date = end_date, 
                         concept_set_name = "CGM",
                         return = "all")
 }

}
