% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_connect.R
\name{aou_connect}
\alias{aou_connect}
\title{Create a connection to the database in All of Us}
\usage{
aou_connect(CDR = getOption("aou.default.cdr"))
}
\arguments{
\item{CDR}{The name of the "curated data repository" that will be used in any
references of the form "{CDR}" or "{cdr}" in the query. Defaults to
getOption("aou.default.cdr"), which is Sys.getenv('WORKSPACE_CDR') if not specified otherwise
(the "mainline" CDR). On the controlled tier, specify the "base" CDR with
`CDR = paste0(Sys.getenv('WORKSPACE_CDR'), "_base")`.}
}
\value{
A `BigQueryConnection` object. This object is also saved as getOption("aou.default.con").
}
\description{
Use this function to create a `BigQueryConnection` object.
You can reference this object to connect to the All of Us database and run
SQL code using, e.g., `dbplyr` or `DBI`. A message is printed with the connection
status (successful or not).
}
\examples{
\dontrun{
con <- aou_connect()
# reference the observation table in the database
dplyr::tbl(con, "observation")
# print a list of the tables in the database
DBI::dbListTables(con)
}
}
