% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_connect.R
\name{aou_connect}
\alias{aou_connect}
\title{Connect to big query database in All of Us}
\usage{
aou_connect(bucket_name = "bucket")
}
\arguments{
\item{bucket_name}{variable name for your bucket. Recommend leaving the default.}
}
\description{
To use, simply run the function without any arguments. It will
print a message if you connect successfully. It will also assign your bucket
to an object in your R environment.
}
