% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_temp_tables.R
\name{aou_create_temp_table}
\alias{aou_create_temp_table}
\title{Creates a temporary table from a local data frame or tibble}
\usage{
aou_create_temp_table(
  df,
  nchar_batch = 1e+06,
  con = getOption("aou.default.con")
)
}
\arguments{
\item{df}{a local dataframe or tibble}

\item{nchar_batch}{approximate number of characters per sql query}

\item{con}{connection}
}
\value{
a reference to a temporary table in the database with the data from \code{df}
}
\description{
Generates a temporary table from a local data frame.
Tables generally will need to be small in size (<1500 rows).
The temporary table will only exist for the current session and will need to be created again a new session.
in a new session.
}
\examples{
\dontshow{if (on_workbench()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
con <- aou_connect()
df <- data.frame(concept_id = c(439331, 4290245, 42535816, 46269813,
                 2784565, 45765502, 434112, 4128031, 435640, 45876808),
                 category = c("AB", "DELIV", "DELIV", "SA", "DELIV",
                 "LB", "DELIV", "DELIV", "PREG", "SA"),
                 gest_value = c(NA, NA, NA, NA, NA, NA, NA, NA, 25, NA))
tmp_tbl = aou_create_temp_table(df)


\dontshow{\}) # examplesIf}
}
