% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_obs_period.R
\name{aou_observation_period}
\alias{aou_observation_period}
\title{Generate an observation period table based on OMOP Conventions}
\usage{
aou_observation_period(
  cohort = NULL,
  persistence_window = 548,
  end_date_buffer = 60,
  exclude_aou_visits = FALSE,
  collect = FALSE
)
}
\arguments{
\item{cohort}{query to a cohort or local dataframe with column "person_id"}

\item{persistence_window}{longest allowable time between visits for the same observation period. defaults to 548 see details}

\item{end_date_buffer}{number of days to add to end date. defaults to 60. see details}

\item{exclude_aou_visits}{whether to exclude All of Us clinical visits (not part of the participants typical EHR) from the observation period}

\item{collect}{whether to collect the data or keep as SQL query}
}
\value{
a sql query or local data frame
}
\description{
Generate an observation period table based on OMOP Conventions
}
\details{
Follows conventions described here: https://ohdsi.github.io/CommonDataModel/ehrObsPeriods.html
}
\examples{
\dontrun{
cohort <-
  dplyr::tbl(con, "cb_search_person") \%>\%
  dplyr::filter(has_ehr_data == 1) \%>\%
  head(100) \%>\%
  dplyr::select(person_id)

observation_periods <- aou_observation_period(cohort,
  persistence_window = 548,
  end_date_buffer = 60,
  collect = FALSE
)
}

}
