% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{aou_survey}
\alias{aou_survey}
\title{Function to query allofus observation table for survey responses}
\usage{
aou_survey(
  cohort,
  questions,
  question_output = "text",
  answer_output = "text",
  con = getOption("aou.default.con"),
  collect = TRUE
)
}
\arguments{
\item{cohort}{tbl or dataframe with a cohort that includes a column called person_id}

\item{questions}{either a vector of concept_ids or concept_codes for questions to return results}

\item{question_output}{how to name the columns. Options include text format ("text"), as concept ids preceded by
"x_" ("concept_id"), or using a custom vector of column names matching the vector of `questions`. Defaults to "text".}

\item{answer_output}{whether to return the survey responses in their text (concept code) format ("text") or as the concept id
for a given response ("concept_id"). Defaults to "text".}

\item{con}{connection to the allofus SQL database. Defaults to getOption("aou.default.con"), which is created automatically with `aou_connect()`}

\item{collect}{whether to return the results as a local (TRUE) or database table}
}
\description{
Because responses to the survey questions stored in the ds_survey table do not
include skipped questions (i.e., missing data!), this function makes it easier to
query the observation table for responses to survey questions so that the skipped
responses are included.

The function will return a dataframe or SQL tbl with the initial cohort table along
with a column for each question included in `questions` and answers for
each person_id in the cells. The column names (questions) and answers can
be returned as the concept_code or concept_id. For each question, a column with
the suffix "_date" is included with the date on which the question was answered.

To find the desired survey questions, use the all of us data dictionary,
survey codebook, athena, data browser, or the allofus R package
modified codebook which can be found here: https://roux-ohdsi.github.io/allofus/articles/searchable_codebook.html

Users may want to do some post-processing. For example PMI_skip or 903096
could be marked as NA.
}
\examples{
\dontrun{
con <- aou_connect()
cohort <- tbl(con, "person") \%>\%
  filter(person_id > 5000000) \%>\%
  select(person_id, year_of_birth, gender_concept_id)
aou_survey(
  cohort = cohort,
  questions = c(1585375, 1586135),
  question_output = "text",
  answer_output = "text"
)
aou_survey(cohort, questions = 1585811, question_output = "pregnancy")
aou_survey(cohort, questions = 1585811, question_output = "text")
aou_survey(cohort, questions = 1585811, question_output = "concept_id")
aou_survey(cohort, questions = c(1585811, 1585386), question_output = c("pregnancy", "insurance"))
aou_survey(cohort, questions = c(1585811, 1585386), question_output = "text")
aou_survey(cohort, questions = c(1585811, 1585386), question_output = "concept_id")
}
}
