% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aou_table_info}
\alias{aou_table_info}
\title{Table of tables, columns, and use for researchers from the CT data dictionary
\itemize{
   \item{\href{https://github.com/roux-ohdsi/allofus/blob/main/data-raw/more_data.R}{Code to generate table}}
}}
\format{
`health_history_codebook`
A data frame with 1685  rows and 9 columns:
\describe{
  \item{table_name}{chr; name of the table}
  \item{columns}{chr; columns in the table}
  \item{recommended_for_research}{chr; whether the table is recomended for research}
}
}
\usage{
aou_table_info
}
\description{
Table of tables, columns, and use for researchers from the CT data dictionary
\itemize{
   \item{\href{https://github.com/roux-ohdsi/allofus/blob/main/data-raw/more_data.R}{Code to generate table}}
}
}
\keyword{datasets}
