% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_connect.R
\name{aou_tables}
\alias{aou_tables}
\title{List tables in the AoU Database}
\usage{
aou_tables(con = getOption("aou.default.con"), remove_NA = TRUE)
}
\arguments{
\item{con}{connection to the database}

\item{remove_NA}{whether to remove tables that are not in the data dictionary}
}
\value{
a dataframe with the table names and the number of columns
}
\description{
List tables in the AoU Database
}
\examples{
\dontrun{
con <- aou_connect()
aou_tables()
}
}
