% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_bucket_to_workspace.R
\name{aou_workspace_to_bucket}
\alias{aou_workspace_to_bucket}
\title{Save a file from your workspace to your bucket.}
\usage{
aou_workspace_to_bucket(files, bucket_name = Sys.getenv("WORKSPACE_BUCKET"))
}
\arguments{
\item{files}{name of file to save}

\item{bucket_name}{name of your bucket. Recommend leaving the default}
}
\description{
This step permanently saves a file you have saved in your workspace
to your bucket where you can always retrieve it. To use, first you need to save the desired
r object as a file (e.g., write.csv(object, filename.csv)) and then run this function
(e.g., aou_workspace_to_bucket(files = "filename.csv")).
}
