% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_functions.R
\name{DATE_DIFF}
\alias{DATE_DIFF}
\title{Date add function to be passed along to SQL}
\usage{
DATE_DIFF(...)
}
\arguments{
\item{...}{additional arguments passed to dbplyr SQL}
}
\value{
Internal, passed along to dbplyr SQL.
}
\description{
Date add function to be passed along to SQL
}
\keyword{internal}
